refactor(studio): extract wrapRoute handlers from server.ts
Shatter activity, brand, and signals wrapRoute handlers into
server/routes/* modules. Drop duplicate rate-limit checks inside
wrapRoute so only the fetch perimeter counts mutating quota.
